+++ applications/GeoMechanicsApplication/test_setup_utilities/model_setup_utilities.cpp
@@ -16,7 +16,7 @@
 #include "custom_elements/plane_strain_stress_state.h"
 #include "custom_elements/small_strain_U_Pw_diff_order_element.hpp"
 #include "custom_elements/three_dimensional_stress_state.h"
-#include "element_setup_utilities.h"
+#include "element_setup_utilities.hpp"
 #include "geometries/tetrahedra_3d_10.h"
 #include "geometries/tetrahedra_3d_4.h"
 #include "geometries/triangle_2d_3.h"
