+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_conditions/test_line_load_2D_diff_order_condition.cpp
@@ -17,7 +17,7 @@
 #include "geometries/line_2d_2.h"
 #include "geometries/line_2d_3.h"
 #include "includes/stream_serializer.h"
-#include "test_setup_utilities/element_setup_utilities.h"
+#include "test_setup_utilities/element_setup_utilities.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 
 #include <string>
