+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_elements/test_axisymmetric_integration_coefficients.cpp
@@ -13,7 +13,7 @@
 //
 
 #include "custom_elements/integration_coefficient_modifier_for_axisymmetric_element.h"
-#include "test_setup_utilities/element_setup_utilities.h"
+#include "test_setup_utilities/element_setup_utilities.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/test_utilities.h"
