+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_elements/test_interface_element.cpp
@@ -22,7 +22,7 @@
 #include "custom_geometries/interface_geometry.hpp"
 #include "custom_utilities/ublas_utilities.h"
 #include "geo_mechanics_application_variables.h"
-#include "test_setup_utilities/element_setup_utilities.h"
+#include "test_setup_utilities/element_setup_utilities.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/test_utilities.h"
