+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_processes/test_find_neighbour_elements_of_conditions.cpp
@@ -11,7 +11,7 @@
 //
 
 #include "custom_processes/find_neighbour_elements_of_conditions_process.h"
-#include "test_setup_utilities/element_setup_utilities.h"
+#include "test_setup_utilities/element_setup_utilities.hpp"
 #include "test_setup_utilities/model_setup_utilities.h"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
