+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_processes/test_find_neighbours_of_interfaces_process.cpp
@@ -13,7 +13,7 @@
 #include "containers/model.h"
 #include "custom_processes/find_neighbours_of_interfaces_process.h"
 #include "includes/kratos_parameters.h"
-#include "test_setup_utilities/element_setup_utilities.h"
+#include "test_setup_utilities/element_setup_utilities.hpp"
 #include "test_setup_utilities/model_setup_utilities.h"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
