+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_utilities/test_extrapolation_utilities.cpp
@@ -15,7 +15,7 @@
 #include "custom_utilities/nodal_extrapolator.h"
 #include "custom_utilities/ublas_utilities.h"
 #include "includes/checks.h"
-#include "test_setup_utilities/element_setup_utilities.h"
+#include "test_setup_utilities/element_setup_utilities.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/test_utilities.h"
