+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_utilities/test_linear_nodal_extrapolator.cpp
@@ -23,7 +23,7 @@
 #include "geometries/tetrahedra_3d_4.h"
 #include "geometries/triangle_2d_3.h"
 #include "geometries/triangle_2d_6.h"
-#include "test_setup_utilities/element_setup_utilities.h"
+#include "test_setup_utilities/element_setup_utilities.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/test_utilities.h"
